detect if the user's future is inconsistent with the results
A framework might wrap the future for adaption, such as Quarkus'
context propagation, which must abide by the future's contract. That
was not the case for the isDone() and dependent actions (whenComplete)
which were inconsistent with the future. If the dependent action was
successful (provided a value and no error) but the future was in-flight
or failed, then the cache's replacement would not result in the needed
policy metadata updates. Instead for an invalid state the cache will
treat it as a failure, log the error, and remove the mapping. There is
no attempt to notify the listener since the future and value as not
aligned, so we don't know what the proper action should be but to warn
and discard. The logging provides as much as can be safely inferred to
help the user track down what cache detected this contract violation.

https://github.com/quarkusio/quarkus/issues/40852
https://github.com/quarkusio/quarkus/issues/50513