Allow users to read the maximum size without locking (fixes #1897)
Policy.Eviction#getMaximum() and weightedSize() now uses acquire
semantics to read the long fields.

Policy.Eviction#setMaximum() continues to lock because it must modify
recalculate the region sizes and adjust multiple fields. It writes to
the maximum's long field using release semantics.

Cache.cleanUp() and the internal cache calls performs its work under
the lock. It continues to read the long fields using plain semantics.
This is allowed since the monitor provides visibility since it is
written under the lock, so when it was a plain read/write we had the
lock provide the barriers. This is no different, we get to use a
stronger release store to allow for a non-lock read via acquire. Thus
there is no performance impact of the stronger memory visibility for
these important fields.