Fix misplaced `@NonNull` annotation in AsyncCache getIfPresent
While the cache will automatically remove a future that resolves to a
null value, it can only do so after it completes. The future that the
user created is returned, not a derived one that might promote the
null to an exception, as otherwise it could break in numerous ways.
Therefore, we cannot know if the in-flight future will resolve to null
or if the call is made before the mapping was discarded, and so this
annotation was misplaced during the jspecify migration.